fastapi/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
aiomysql
python-dotenv
cryptography
pyjwt
httpx
firebase_admin